% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiPlot.R
\name{epiplot}
\alias{epiplot}
\title{Create Explanatory and Descriptive Plots}
\usage{
epiplot(data, x, y, graph, fill, title, xlab, ylab, legend)
}
\description{
This function plots data according to given variables and parameters from a dataset.
EpiPlot includes the following plots: bar plots, grouped bar plots, stacked bar plots, box plots, box plots with correlating dots, histograms,
density plots, scatter plots, scatter plots with correlating line, and linear regression plots.
}
\details{
Arguments:
data: dataset of choice formatted as specified
x: x variable
y: y variable
graph: specified graph of choice
fill: additional variable used for fill of the plot
title: Title of desired graph
xlab: Label for the x-axis
ylab: Label for the y-axis
Label: Label for graph legend
}
