% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Midwest.R
\name{map_midwest_prevalence}
\alias{map_midwest_prevalence}
\title{Create Midwest Regional Prevalence Map}
\usage{
map_midwest_prevalence(data, existing_cases, population, state, year)
}
\description{
This function maps regional prevalence of a variable of choice across all time points
of the dataset. Midwest includes: Ohio, Michigan, Indiana, Wisconsin, Illinois,
Minnesota, Iowa, Missouri, North Dakota, South Dakota, Nebraska, and Kansas.
}
\details{
Arguments:
data: dataset of choice formatted as specified
existing cases: variable for prevalence calculation
population: variable for population estimate
state: state arugment must be "state_full"
year: year argument must be "year"
}
