% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/National.R
\name{map_national_prevalence}
\alias{map_national_prevalence}
\title{Create National Prevalence Map}
\usage{
map_national_prevalence(data, existing_cases, population, state, year)
}
\description{
This function maps national prevalence of a variable of choice across
   all time points of the dataset.
}
\details{
Arguments:
data: dataset of choice formatted as specified
existing cases: variable for prevalence calculation
population: variable for population estimate
state: state arugment must be "state_full"
year: year argument must be "year"
}
