% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Northeast.R
\name{map_northeast_prevalence}
\alias{map_northeast_prevalence}
\title{Create Northeast Regional Prevalence Map}
\usage{
map_northeast_prevalence(data, existing_cases, population, state, year)
}
\description{
This function maps regional prevalence of a variable of choice across all time points
of the dataset. Northeast includes: Massachusetts, Rhode Island, Connecticut, Maine,
Vermont, New Hampshire, New York, New Jersey, and Pennsylvania.
}
\details{
Arguments:
data: dataset of choice formatted as specified
existing cases: variable for prevalence calculation
population: variable for population estimate
state: state arugment must be "state_full"
year: year argument must be "year"
}
