% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/South.R
\name{map_south_prevalence}
\alias{map_south_prevalence}
\title{Create South Regional Prevalence Map}
\usage{
map_south_prevalence(data, existing_cases, population, state, year)
}
\description{
This function maps regional prevalence of a variable of choice across all time points
of the dataset. South includes: Delaware, Maryland, West Virginia, Virginia, Kentucky,
Tennessee, North Carolina, South Carolina, Georgia, Alabama, Mississippi, Arkansas,
Louisiana, Florida, Texas, and Oklahoma.
}
\details{
Arguments:
data: dataset of choice formatted as specified
existing cases: variable for prevalence calculation
population: variable for population estimate
state: state arugment must be "state_full"
year: year argument must be "year"
}
