% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/West.R
\name{map_west_prevalence}
\alias{map_west_prevalence}
\title{Create West Regional Prevalence Map}
\usage{
map_west_prevalence(data, existing_cases, population, state, year)
}
\description{
This function maps regional prevalence of a variable of choice across all time points
of the dataset. West includes: Montana, Wyoming, Colorado, New Mexico, Arizona,
Utah, California, Idaho, Nevada, Oregon, Washington, Hawaii, and Alaska.
}
\details{
Arguments:
data: dataset of choice formatted as specified
existing cases: variable for prevalence calculation
population: variable for population estimate
state: state arugment must be "state_full"
year: year argument must be "year"
}
